Extracts config_utils file
